Added some notes for people getting here from the installation steps.
So the first time I visited this page I was following a link on the main ZoneMTA github page, so I got a little confused, this tries to clarify this for newcomers.

I also did not understand the wrapper concept, so I'm adding some extra information about what this is and what it does. Please check for accuracy of the terms, especially when talking about the wrapper. This is a concept that was foreign to me and I had to research.